+++ frontend/src/stores/registrationFormStore.ts
@@ -1,18 +1,5 @@
+import { RegistrationFormState } from "@/types";
 import { defineStore } from "pinia";
-
-export interface RegistrationFormState {
-  surname: string;
-  name: string;
-  surnameKana: string;
-  nameKana: string;
-  gender: string;
-  year: number;
-  month: number;
-  date: number;
-  prefecture: string;
-  address: string;
-  email: string;
-}
 
 const defaultFormValue: RegistrationFormState = {
   surname: "",
