+++ frontend/src/types/index.ts
@@ -1,0 +1,13 @@
+export interface RegistrationFormState {
+  surname: string;
+  name: string;
+  surnameKana: string;
+  nameKana: string;
+  gender: string;
+  year: number;
+  month: number;
+  date: number;
+  prefecture: string;
+  address: string;
+  email: string;
+}
